+++ packages/twenty-server/src/modules/calendar-messaging-participant-manager/utils/is-email-blocklisted.util.ts
@@ -1,3 +1,5 @@
+// TODO: Move inside blocklist module
+
 export const isEmailBlocklisted = (
   channelHandle: string,
   email: string | null | undefined,
