+++ packages/twenty-server/src/modules/connected-account/auto-companies-and-contacts-creation/utils/get-domain-name-from-handle.util.ts
